pyproject.toml
[tool.poetry]
name = "flow-rms-invoice-api"
version = "0.1.0"
description = "Flow RMS multi-tenant invoice reconciliation platform"
authors = ["Flow RMS <engineering@flowrms.example>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.110.3"
strawberry-graphql = "^0.287.3"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
SQLAlchemy = "^2.0.36"
alembic = "^1.13.1"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88
combine_as_imports = true

[tool.ruff]
line-length = 88
target-version = "py313"
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
